+++ cmd/set_test.go
@@ -2,7 +2,6 @@ package cmd
 
 import (
 	"bytes"
-	"fmt"
 	"strings"
 	"testing"
 
@@ -14,12 +13,6 @@ const (
 )
 
 func Test_setCommand(t *testing.T) {
-	voidCmd := newSetCommand(
-		&GlobalOption{
-			Out:    &bytes.Buffer{},
-			ErrOut: &bytes.Buffer{},
-		},
-	)
 	tests := []struct {
 		name          string
 		command       string
@@ -63,17 +56,31 @@ func Test_setCommand(t *testing.T) {
 			wantErr:       false,
 		},
 		{
+			name:          "two args encryption and specify KMS key ID",
+			command:       "/service1/dev/key3 value3 --encrypt --kms-key-id key",
+			wantOutWriter: "",
+			wantErrWriter: "",
+			wantErr:       false,
+		},
+		{
+			name:          "two args specify KMS key ID without encryption",
+			command:       "/service1/dev/key3 value3 --kms-key-id key",
+			wantOutWriter: "",
+			wantErrWriter: "Failed to execute PutParameter API.\nKMS Key ID must be used with SecureString type.",
+			wantErr:       true,
+		},
+		{
 			name:          "one arg",
 			command:       "/service1/dev/key1",
-			wantOutWriter: fmt.Sprintf("Error: accepts 2 arg(s), received 1\n%s%s", voidCmd.UsageString(), usageSetHelp),
-			wantErrWriter: "",
+			wantOutWriter: "",
+			wantErrWriter: "accepts 2 arg(s), received 1",
 			wantErr:       true,
 		},
 		{
 			name:          "no args",
 			command:       "",
-			wantOutWriter: fmt.Sprintf("Error: accepts 2 arg(s), received 0\n%s%s", voidCmd.UsageString(), usageSetHelp),
-			wantErrWriter: "",
+			wantOutWriter: "",
+			wantErrWriter: "accepts 2 arg(s), received 0",
 			wantErr:       true,
 		},
 	}
@@ -95,15 +102,19 @@ func Test_setCommand(t *testing.T) {
 				cmd.SetArgs(strings.Split(tt.command, " "))
 			}
 
-			if err := cmd.Execute(); (err != nil) != tt.wantErr {
+			err := cmd.Execute()
+			if (err != nil) != tt.wantErr {
 				t.Errorf("get() error = %v, wantErr %v", err, tt.wantErr)
 				return
 			}
-			if gotOutWriter := outWriter.String(); gotOutWriter != tt.wantOutWriter {
-				t.Errorf("get() = %v, want %v", gotOutWriter, tt.wantOutWriter)
-			}
-			if gotErrWriter := errWriter.String(); gotErrWriter != tt.wantErrWriter {
-				t.Errorf("get() = %v, want %v", gotErrWriter, tt.wantErrWriter)
+			if err != nil {
+				if err.Error() != tt.wantErrWriter {
+					t.Errorf("get() = %v, want %v", err.Error(), tt.wantErrWriter)
+				}
+			} else {
+				if gotOutWriter := outWriter.String(); gotOutWriter != tt.wantOutWriter {
+					t.Errorf("get() = %v, want %v", gotOutWriter, tt.wantOutWriter)
+				}
 			}
 		})
 	}